requirements.txt
flask>=3.1.2
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
